Placed sources according to Maven convention.
