+++ src/main/java/com/runjva/sourceforge/jsocks/protocol/AuthenticationNone.java
