+++ src/main/java/com/runjva/sourceforge/jsocks/protocol/InetRange.java
