+++ src/main/java/com/runjva/sourceforge/jsocks/protocol/Socks4Message.java
