+++ src/main/java/com/runjva/sourceforge/jsocks/protocol/SocksDialog.java
