+++ src/main/java/com/runjva/sourceforge/jsocks/protocol/SocksException.java
