+++ src/main/java/com/runjva/sourceforge/jsocks/server/ServerAuthenticatorNone.java
