+++ src/main/resources/jsocks/main/SocksEcho.gif
